backend/requirements.txt
# ===========================
# ComplyIQ Backend Dependencies (Python 3.12)
# ===========================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.17.0
asyncpg==0.29.0

# Async & Concurrency
celery==5.3.4
redis==5.0.1
aioredis==2.0.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-multipart==0.0.6

# Web Scraping & Browser Automation
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Data Validation & Serialization
python-dotenv==1.0.0
httpx==0.25.2

# API Documentation
swagger-ui==1.1.0

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0
tldextract==5.0.1  # For domain analysis

# AI/LLM Integration (OpenAI)
openai==1.3.9

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Production Server
gunicorn==21.2.0
